+++ api/lib/opentelemetry/baggage/propagation/text_map_propagator.rb
@@ -11,10 +11,15 @@ module OpenTelemetry
     module Propagation
       # Propagates baggage using the W3C Baggage format
       class TextMapPropagator
+        # Maximums according to W3C Baggage spec
+        MAX_ENTRIES = 180
+        MAX_ENTRY_LENGTH = 4096
+        MAX_TOTAL_LENGTH = 8192
+
         BAGGAGE_KEY = 'baggage'
         FIELDS = [BAGGAGE_KEY].freeze
 
-        private_constant :BAGGAGE_KEY, :FIELDS
+        private_constant :BAGGAGE_KEY, :FIELDS, :MAX_ENTRIES, :MAX_ENTRY_LENGTH, :MAX_TOTAL_LENGTH
 
         # Inject in-process baggage into the supplied carrier.
         #
@@ -24,11 +29,12 @@ module OpenTelemetry
         #   will be used to write context into the carrier, otherwise the default
         #   text map setter will be used.
         def inject(carrier, context: Context.current, setter: Context::Propagation.text_map_setter)
-          baggage = context[ContextKeys.baggage_key]
+          baggage = OpenTelemetry.baggage.raw_entries(context: context)
 
           return if baggage.nil? || baggage.empty?
 
-          setter.set(carrier, BAGGAGE_KEY, encode(baggage))
+          encoded_baggage = encode(baggage)
+          setter.set(carrier, BAGGAGE_KEY, encoded_baggage) unless encoded_baggage&.empty?
           nil
         end
 
@@ -49,16 +55,16 @@ module OpenTelemetry
 
           entries = header.gsub(/\s/, '').split(',')
 
-          baggage = entries.each_with_object({}) do |entry, memo|
-            # The ignored variable below holds properties as per the W3C spec.
-            # OTel is not using them currently, but they might be used for
-            # metadata in the future
-            kv, = entry.split(';', 2)
-            k, v = kv.split('=').map!(&CGI.method(:unescape))
-            memo[k] = v
+          OpenTelemetry.baggage.build(context: context) do |builder|
+            entries.each do |entry|
+              # Note metadata is currently unused in OpenTelemetry, but is part
+              # the W3C spec where it's referred to as properties. We preserve
+              # the properties (as-is) so that they can be propagated elsewhere.
+              kv, meta = entry.split(';', 2)
+              k, v = kv.split('=').map!(&CGI.method(:unescape))
+              builder.set_value(k, v, metadata: meta)
+            end
           end
-
-          context.set_value(ContextKeys.baggage_key, baggage)
         rescue StandardError => e
           OpenTelemetry.logger.debug "Error extracting W3C baggage: #{e.message}"
           context
@@ -75,9 +81,27 @@ module OpenTelemetry
         private
 
         def encode(baggage)
-          baggage.inject(+'') do |memo, (k, v)|
-            memo << CGI.escape(k.to_s) << '=' << CGI.escape(v.to_s) << ','
-          end.chop!
+          result = +''
+          encoded_count = 0
+          baggage.each_pair do |key, entry|
+            break unless encoded_count < MAX_ENTRIES
+
+            encoded_entry = encode_value(key, entry)
+            next unless encoded_entry.size <= MAX_ENTRY_LENGTH &&
+                        encoded_entry.size + result.size <= MAX_TOTAL_LENGTH
+
+            result << encoded_entry << ','
+            encoded_count += 1
+          end
+          result.chop!
+        end
+
+        def encode_value(key, entry)
+          result = +"#{CGI.escape(key.to_s)}=#{CGI.escape(entry.value.to_s)}"
+          # We preserve metadata recieved on extract and assume it's already formatted
+          # for transport. It's sent as-is without further processing.
+          result << ";#{entry.metadata}" if entry.metadata
+          result
         end
       end
     end
